revert: undo all Phase 2A old-plan merges
Reverts 6 PRs (#393–#398) that were implemented against the old Phase 2A
requirements before the plan was restructured. The new plan (F1 Demo Harness
+ Real Data Pipeline) has different scope and feature decomposition.

Reverted commits:
- b1d022b ci(workspace): add workspace E2E workflow [VASTU-2A-211] (#398)
- c1e5d0b feat(workspace): add chart keyboard accessibility [VASTU-2A-210] (#397)
- 6b9670d feat(workspace): custom cell formatter registry [VASTU-2A-205] (#396)
- c20a969 feat(shared,shell): data engine types, translators, and query API [VASTU-2A-202] (#395)
- c1bafab feat(workspace): add application page registry [VASTU-2A-206] (#394)
- cc8917b feat(shared): add Prisma schema composition infrastructure [VASTU-2A-201] (#393)
