+++ packages/workspace/messages/en.json
@@ -228,13 +228,6 @@
   "chart.config.scaleType": "Y-axis scale",
   "chart.config.scale.linear": "Linear",
   "chart.config.scale.log": "Log",
-  "chart.a11y.navHint": "Use arrow keys to navigate data points. Press Enter to drill down.",
-  "chart.a11y.announcement": "{type} chart. {series}: {x}, {value}. Point {index} of {total}.",
-  "chart.a11y.noValue": "no data",
-  "chart.a11y.viewAsTable": "View as table",
-  "chart.a11y.viewAsChart": "View as chart",
-  "chart.a11y.table.series": "Series",
-  "chart.a11y.table.value": "Value",
   "drawer.ariaLabel": "Record detail",
   "drawer.close": "Close",
   "drawer.loading.ariaLabel": "Loading record",
