+++ packages/workspace/src/components/SidebarNav/index.ts
@@ -3,3 +3,5 @@ export { SidebarSection } from './SidebarSection';
 export { SidebarItem } from './SidebarItem';
 export { SidebarSearch } from './SidebarSearch';
 export { SidebarUserAvatar } from './SidebarUserAvatar';
+export { MOCK_PAGES } from './mockPages';
+export type { MockPage } from './mockPages';
