--- packages/workspace/src/components/VastuChart/ChartAccessibleTable.module.css
@@ -1,113 +1,0 @@
-/**
- * ChartAccessibleTable styles.
- *
- * All colors via --v-* CSS custom properties.
- * No hardcoded hex values.
- *
- * Implements US-210 AC-6 (VASTU-2A-210c).
- */
-
-/* ─── Wrapper ────────────────────────────────────────────────────────────── */
-
-.wrapper {
-  width: 100%;
-  overflow-x: auto;
-}
-
-/* ─── Table ──────────────────────────────────────────────────────────────── */
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  font-family: var(--v-font-sans);
-  font-size: 12px; /* --v-text-sm */
-  font-weight: 400;
-  color: var(--v-text-primary);
-}
-
-/* ─── Header ─────────────────────────────────────────────────────────────── */
-
-.table thead tr {
-  background-color: var(--v-bg-secondary);
-  border-bottom: 1px solid var(--v-border-default);
-}
-
-.table th {
-  padding: 6px 12px;
-  text-align: left;
-  font-weight: 500; /* --v-font-medium */
-  font-size: 11px; /* --v-text-xs */
-  color: var(--v-text-secondary);
-  white-space: nowrap;
-  border-right: 1px solid var(--v-border-subtle);
-  user-select: none;
-}
-
-.table th:last-child {
-  border-right: none;
-}
-
-/* ─── Body rows ──────────────────────────────────────────────────────────── */
-
-.table tbody tr {
-  border-bottom: 1px solid var(--v-border-subtle);
-  transition: background-color var(--v-duration-fast) var(--v-ease-default);
-}
-
-.table tbody tr:last-child {
-  border-bottom: none;
-}
-
-.table tbody tr:hover {
-  background-color: var(--v-bg-tertiary);
-}
-
-.table td {
-  padding: 4px 12px; /* compact density */
-  vertical-align: middle;
-  border-right: 1px solid var(--v-border-subtle);
-}
-
-.table td:last-child {
-  border-right: none;
-}
-
-/* ─── Series swatch ──────────────────────────────────────────────────────── */
-
-.seriesSwatch {
-  display: inline-block;
-  width: 8px;
-  height: 8px;
-  border-radius: 2px; /* --v-radius-sm */
-  flex-shrink: 0;
-  vertical-align: middle;
-  margin-right: var(--v-space-1);
-}
-
-/* ─── Null value ─────────────────────────────────────────────────────────── */
-
-.nullValue {
-  color: var(--v-text-tertiary);
-}
-
-/* ─── Screen reader only ─────────────────────────────────────────────────── */
-
-.srOnly {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
-
-/* ─── Reduced motion ─────────────────────────────────────────────────────── */
-
-@media (prefers-reduced-motion: reduce) {
-  .table tbody tr {
-    transition: none;
-  }
-}
