--- packages/workspace/src/components/VastuChart/ChartKeyboardNav.module.css
@@ -1,45 +1,0 @@
-/**
- * ChartKeyboardNav styles.
- *
- * All colors via --v-* CSS custom properties.
- * No hardcoded hex values.
- *
- * Implements US-210 AC-1 (VASTU-2A-210a).
- */
-
-/* ─── Container ──────────────────────────────────────────────────────────── */
-
-.container {
-  position: relative;
-  width: 100%;
-  outline: none;
-}
-
-/* Focus ring: visible only on keyboard navigation, never on mouse click. */
-.container:focus-visible {
-  outline: none;
-  box-shadow: var(--v-shadow-focus);
-  border-radius: var(--v-radius-md);
-}
-
-/* ─── Screen reader live region ──────────────────────────────────────────── */
-
-.srLiveRegion {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  padding: 0;
-  margin: -1px;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  white-space: nowrap;
-  border: 0;
-}
-
-/* ─── Reduced motion ─────────────────────────────────────────────────────── */
-
-@media (prefers-reduced-motion: reduce) {
-  .container {
-    transition: none;
-  }
-}
