+++ packages/workspace/src/components/VastuChart/ChartTooltip.tsx
@@ -10,12 +10,7 @@
  * - Square color swatch per series (10×10px, --v-radius-sm)
  * - Values sorted by value descending
  *
- * Value display:
- *   1. If a seriesConfig provides a `displayType`, the registered formatter's
- *      render function is used to format the value (VASTU-2A-205c).
- *   2. Otherwise, the raw value is displayed as a string.
- *
- * Implements US-135 AC-5, extended by VASTU-2A-205c.
+ * Implements US-135 AC-5.
  *
  * Note: This component is passed as a render function to Recharts <Tooltip content={...}>.
  * Recharts injects active, payload, label at render time.
@@ -23,7 +18,6 @@
 
 import React from 'react';
 import type { SeriesConfig } from './types';
-import { getFormatter } from '../../formatters/registry';
 
 export interface ChartTooltipProps {
   /** Injected by Recharts: whether the tooltip is active. */
@@ -43,28 +37,6 @@ export interface ChartTooltipProps {
   resolvedColors: Record<string, string>;
 }
 
-/**
- * Format a single tooltip value using the formatter registry when a displayType
- * is configured on the matching SeriesConfig, otherwise falls back to String().
- */
-function formatTooltipValue(
-  value: string | number | null | undefined,
-  seriesConfig: SeriesConfig | undefined,
-): React.ReactNode {
-  if (value == null) return '\u2014';
-
-  const displayType = seriesConfig?.displayType;
-
-  if (displayType) {
-    const formatter = getFormatter(displayType);
-    if (formatter) {
-      return formatter.render({ value, row: {} });
-    }
-  }
-
-  return String(value);
-}
-
 export function ChartTooltip({ active, payload, label, series, resolvedColors }: ChartTooltipProps) {
   if (!active || !payload || payload.length === 0) {
     return null;
@@ -80,7 +52,6 @@ export function ChartTooltip({ active, payload, label, series, resolvedColors }:
         name: seriesConfig?.name ?? String(entry.name ?? ''),
         value: entry.value as string | number | null | undefined,
         color: (dKey && resolvedColors[dKey]) ? resolvedColors[dKey] : (String(entry.color ?? 'var(--v-text-tertiary)')),
-        seriesConfig,
       };
     })
     // Sort by value descending (nulls last)
@@ -107,7 +78,7 @@ export function ChartTooltip({ active, payload, label, series, resolvedColors }:
             />
             <span style={tooltipStyles.name}>{entry.name}</span>
             <span style={tooltipStyles.value}>
-              {formatTooltipValue(entry.value, entry.seriesConfig)}
+              {entry.value != null ? String(entry.value) : '\u2014'}
             </span>
           </div>
         ))}
