+++ packages/workspace/src/components/VastuChart/index.ts
@@ -9,17 +9,6 @@ export { ChartLegend } from './ChartLegend';
 export { ChartTooltip } from './ChartTooltip';
 export type { ChartTooltipProps } from './ChartTooltip';
 export { ChartConfigPanel } from './ChartConfigPanel';
-export { ChartKeyboardNav } from './ChartKeyboardNav';
-export type { ChartKeyboardNavProps } from './ChartKeyboardNav';
-export { ChartAccessibleTable } from './ChartAccessibleTable';
-export type { ChartAccessibleTableProps } from './ChartAccessibleTable';
-export { useChartKeyboardNav } from './useChartKeyboardNav';
-export type {
-  ChartKeyboardNavState,
-  ChartKeyboardNavHandlers,
-  UseChartKeyboardNavOptions,
-  UseChartKeyboardNavResult,
-} from './useChartKeyboardNav';
 export { CHART_SERIES_COLORS, CHART_OTHER_COLOR, getSeriesColor } from './chartColors';
 export type {
   ChartType,
