+++ packages/workspace/src/index.ts
@@ -31,18 +31,6 @@ export {
 } from './panels/registry';
 export { WelcomePanel, WELCOME_PANEL_TYPE_ID } from './panels/WelcomePanel';
 
-// Page registry (VASTU-2A-206)
-export {
-  registerPage,
-  getAllPages,
-  getPageById,
-  mergeWithDynamic,
-  unregisterPage,
-  clearPageRegistry,
-} from './pages/registry';
-export { registerPagePanels } from './panels/index';
-export type { PageDefinition, MergedPage, PageSection } from './pages/types';
-
 // Stores
 export { useSidebarStore } from './stores/sidebarStore';
 export { usePanelStore, openPanelByTypeId } from './stores/panelStore';
@@ -346,23 +334,3 @@ export type {
   AlertCardDef,
   DashboardChartDataPoint,
 } from './stores/dashboardStore';
-
-// FormatterRegistry (VASTU-2A-205)
-export {
-  registerFormatter,
-  getFormatter,
-  getAllFormatters,
-  hasFormatter,
-  unregisterFormatter,
-  clearFormatterRegistry,
-  BUILTIN_FORMATTERS,
-  registerBuiltinFormatters,
-} from './formatters';
-export type {
-  FormatterDefinition,
-  FormatterMeta,
-  FormatterRenderContext,
-  FormatterSortContext,
-  FormatterExportContext,
-  FormatterFilterContext,
-} from './formatters';
